Fixed meter-indexing problem
RpermMeter is now a data structure containing a Vector of Rperms and a
member storing the length of that vector (in other words, it knows its
own length). When we access the RpermMeter, we take the modulo of the
random index generated by @Fortuna@ (now in a range 0-5) and the
rpermMax stored in the rpermMeter so we always get a result.

We also created dedicated buildVpermTable and buildRpermTable functions
to allow for this and make the Ark data modules clearer.